requirements.txt
streamlit==1.26.0
Pillow==9.4.0

